Ai-Model-Working/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
tensorflow>=2.16.1,<2.21.0
numpy>=1.24.0,<2.0.0
opencv-python>=4.8.0
requests>=2.31.0
pydantic>=2.5.0
python-dotenv>=1.0.0

